backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
numpy
sentence-transformers
pinecone-client
langchain
langchain-pinecone
langchain-google-genai
scikit-learn
umap-learn
Pillow
torch
torchvision
matplotlib
seaborn
transformers
pydantic-settings
